Add standardizer and preprocessor classes, restructure
